requirements.txt
tensorflow
numpy
matplotlib
Pillow
scipy
tk 
pygame
pyserial